break up handlers module and fix imports
